etc: maindeco move to make/maindeco
