--- front/src/pages/MainDeco/index.tsx
@@ -1,1 +1,0 @@
-export {default as MainDeco} from './MainDeco';
