+++ front/src/pages/Make/MainDeco/DecoEnroll.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef } from 'react';
 import { NavigateFunction, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import theme from '../../utils/theme';
+import theme from '../../../utils/theme';
 
 interface NaviProps {
   visible: [number, React.Dispatch<React.SetStateAction<number>>];
@@ -128,13 +128,10 @@ const DecoEnroll = (props: NaviProps) => {
   const closeRef = useRef<HTMLDivElement>(null);
   const navigate = useNavigate();
 
-
   return (
     <>
       <StyledBody
-        onClick={() =>
-          CloseNav(props, closeRef, setIsFocus, navigate, 'close')
-        }
+        onClick={() => CloseNav(props, closeRef, setIsFocus, navigate, 'close')}
       />
 
       {isFocus ? (
