+++ front/src/pages/Make/MainDeco/DecoProvider.tsx
@@ -1,8 +1,8 @@
 import React, { useState, createContext } from 'react';
 
 interface DecoContextType {
-  decoID: number;
-  setDecoID: React.Dispatch<React.SetStateAction<number>>;
+  mainDecoID: number;
+  setMainDecoID: React.Dispatch<React.SetStateAction<number>>;
   color: string;
   setColor: React.Dispatch<React.SetStateAction<string>>;
   letterID: number;
@@ -14,8 +14,8 @@ interface DecoContextType {
 }
 
 const DecoContext = createContext<DecoContextType>({
-  decoID: 0,
-  setDecoID: () => {},
+  mainDecoID: 0,
+  setMainDecoID: () => {},
   color: '#ff0000',
   setColor: () => {},
   letterID: 0,
@@ -29,16 +29,17 @@ const DecoContext = createContext<DecoContextType>({
 const DecoProvider: React.FC<{ children: React.ReactNode }> = ({
   children
 }) => {
-  const [decoID, setDecoID] = useState<number>(0);
+  const [mainDecoID, setMainDecoID] = useState<number>(0);
   const [color, setColor] = useState<string>('#ff0000');
   const [letterID, setLetterID] = useState<number>(0);
   const [content, setContent] = useState<string>('');
   const [sender, setSender] = useState<string>('');
+
   return (
     <DecoContext.Provider
       value={{
-        decoID,
-        setDecoID,
+        mainDecoID,
+        setMainDecoID,
         color,
         setColor,
         letterID,
