+++ front/src/pages/Make/MainDeco/MainDeco.tsx
@@ -1,5 +1,5 @@
-import { UIContainer } from '../../components';
-import Steps from '../MainDeco/Steps';
+import { UIContainer } from '../../../components';
+import Steps from './Steps';
 import { SnowballNameProvider } from './SnowballNameProvider';
 
 const MainDeco = () => {
