+++ front/src/pages/Make/MainDeco/MakeSnowballCanvas.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Canvas } from '@react-three/fiber';
 import { OrbitControls } from '@react-three/drei';
 import styled from 'styled-components';
