+++ front/src/pages/Make/MainDeco/index.tsx
@@ -1,0 +1,1 @@
+export { default as MainDeco } from './MainDeco';
