+++ front/src/pages/Make/index.tsx
@@ -1,2 +1,3 @@
 export { Nickname } from './Nickname';
 export { Snowball } from './Snowball';
+export { MainDeco } from './MainDeco';
