+++ front/src/pages/index.tsx
@@ -1,6 +1,5 @@
 export { Intro } from './Intro';
 export { Visit, Deco } from './Visit';
-export { Nickname, Snowball } from './Make';
+export { Nickname, Snowball, MainDeco } from './Make';
 export { Main } from './Main';
 export { Wrong } from './Wrong';
-export { MainDeco } from './MainDeco';
